requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
PyJWT==2.10.1
pytest==8.3.5
pytest-flask==1.3.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
python-dotenv
